backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
django-filter>=24.1
channels>=4.0
channels-redis>=4.2
daphne>=4.1
python-dotenv>=1.0
requests>=2.31
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
httpx>=0.27
Pillow>=10.2
cryptography>=42.0
python-dateutil>=2.8
